Main refactor for upcoming azure integration -SearchResult is now part of util -Moved util.webSocket in webSocket
