+++ src/main/java/it/polpetta/libris/IQuery.java
@@ -1,7 +1,7 @@
 package it.polpetta.libris;
 
 
-import it.polpetta.libris.google.imageSearch.SearchResult;
+import it.polpetta.libris.utils.SearchResult;
 
 /**
  * Created by davide on 28/04/17.
