+++ src/main/java/it/polpetta/libris/google/imageSearch/searchers/ISearcher.java
@@ -1,6 +1,6 @@
 package it.polpetta.libris.google.imageSearch.searchers;
 
-import it.polpetta.libris.google.imageSearch.SearchResult;
+import it.polpetta.libris.utils.SearchResult;
 
 import java.io.IOException;
