+++ src/main/java/it/polpetta/libris/google/imageSearch/searchers/LocalPhotoSearcher.java
@@ -1,7 +1,7 @@
 package it.polpetta.libris.google.imageSearch.searchers;
 
 import it.polpetta.libris.google.imageSearch.Coordinates;
-import it.polpetta.libris.google.imageSearch.SearchResult;
+import it.polpetta.libris.utils.SearchResult;
 
 import java.io.IOException;
 import java.net.URI;
